fix some minor warnings
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@883 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ kicker.cpp
@@ -567,7 +567,7 @@ STDMETHODIMP Kicker::InterfaceSupportsErrorInfo(REFIID riid)
 		&IID_IKicker,
 	};
 
-	for (int i=0;i<sizeof(arr)/sizeof(arr[0]);i++)
+	for (size_t i=0;i<sizeof(arr)/sizeof(arr[0]);i++)
 	{
 		if (InlineIsEqualGUID(*arr[i],riid))
 			return S_OK;
