+++ wintimer.cpp
@@ -12,15 +12,15 @@ void wintimer_init()
 	QueryPerformanceCounter( &sTimerStart );
 }
 
-LONGLONG usec()
+unsigned long long usec()
 {
 	if( sTimerInit == 0 ) return 0;
 
     LARGE_INTEGER TimerNow;
 	QueryPerformanceCounter( &TimerNow );
-	const LONGLONG cur_tick = TimerNow.QuadPart - sTimerStart.QuadPart;
-	return (TimerFreq.QuadPart < 100000000ull) ? (LONGLONG)((unsigned long long)cur_tick*1000000ull/(unsigned long long)TimerFreq.QuadPart)
-		                                       : (LONGLONG)((unsigned long long)cur_tick*1000ull/((unsigned long long)TimerFreq.QuadPart/1000ull));
+	const unsigned long long cur_tick = (unsigned long long)(TimerNow.QuadPart - sTimerStart.QuadPart);
+	return ((unsigned long long)TimerFreq.QuadPart < 100000000ull) ? (cur_tick*1000000ull/(unsigned long long)TimerFreq.QuadPart)
+		                                                           : (cur_tick*1000ull/((unsigned long long)TimerFreq.QuadPart/1000ull));
 }
 
 U32 msec()
